backend/services/core_api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic>=2.3.0
pydantic-settings>=2.0.0
asyncpg==0.28.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
clerk-sdk-python==0.1.0
python-dotenv==1.0.0
pyjwt==2.8.0
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.27.0
